Linear Regression Example (#10)
* Linear Regression Example

This simple example shows how to use the package to sample from the posterior distribution of a static Bayesian model and to estimate the model evidence. The example was originally presented in the literature as a model choice problem which is simple enough for the true log evidence estimates (and Bayes factor) to be obtainable via numerical integration. Users can select which model they wish to perform inference on (model 1 or 2).

* Minor documentation changes

Giving more detail about the two competing models for the radiata pine regression example and removing obsolete comments about data handling.